refactor: Move Container_Factory to core directory with proper namespace
- Move Container_Factory from includes/ to includes/core/ directory
- Update namespace from WcAiReviewResponder to WcAiReviewResponder\Core
- Update import in main plugin file to use new namespace
- Improve directory organization by separating infrastructure from business logic

Benefits:
- Clear separation between core infrastructure and domain logic
- Better organization for future core services (config, logging, etc.)
- Follows common patterns for infrastructure layer organization
- Maintains PSR-4 compliance with namespace-to-directory mapping
